Move ValueHelper to Support, update README
