+++ resources/views/components/form/feedback.blade.php
@@ -9,7 +9,7 @@
     'showSubErrors' => false,
 ])
 @php
-    use Portavice\Bladestrap\Components\Helpers\ValueHelper;
+    use Portavice\Bladestrap\Support\ValueHelper;
 
     $attributes = $attributes->class([
         config('bladestrap.classes.form.invalid-feedback'),
