+++ resources/views/components/form/field.blade.php
@@ -50,8 +50,8 @@
     'required' => false,
 ])
 @php
-    use Portavice\Bladestrap\Components\Helpers\ValueHelper;
     use Portavice\Bladestrap\Support\OptionCollection;
+    use Portavice\Bladestrap\Support\ValueHelper;
 
     /** @var \Illuminate\View\ComponentAttributeBag $attributes */
     $containerAttributes = $attributes->filterAndTransform('container-');
